+++ main.cc
@@ -73,7 +73,7 @@ c_rnd<double>	dbl_rnd;
 
 void init() {
 	prnt = printf;
-	current = (int)&data;
+	current = (long)&data;
 
 	module = dopen( NULL );
 
@@ -86,82 +86,82 @@ void init() {
 	data.buffer = *((int*)&pi);
 	data.type = 34;
 	data.length = 0;
-	name.buffer = (int)"PI";
+	name.buffer = (long)"PI";
 	name.type = 8;
 	name.length = 2;
 	store( name );
 
 	data.buffer = *((int*)&e);
-	name.buffer = (int)"e";
+	name.buffer = (long)"e";
 	name.length = 1;
 	store( name );
 
 	data.buffer = 1;
 	data.type = 32;
-	name.buffer = (int)"one";
+	name.buffer = (long)"one";
 	name.length = 3;
 	store( name );
 
 	data.buffer = 0;
-	name.buffer = (int)"nil";
+	name.buffer = (long)"nil";
 	name.length = 3;
 	store( name );
 
 	data.buffer = sizeof(double)*8+2;
-	name.buffer = (int)"double";
+	name.buffer = (long)"double";
 	name.length = strlen((char*)name.buffer);
 	store( name );
 
 	data.buffer = sizeof(float)*8+2;
-	name.buffer = (int)"float";
+	name.buffer = (long)"float";
 	name.length = strlen((char*)name.buffer);
 	store( name );
 
-	data.buffer = sizeof(int)*8;
-	name.buffer = (int)"int";
+	data.buffer = sizeof(long)*8;
+	name.buffer = (long)"int";
 	name.length = strlen((char*)name.buffer);
 	store( name );
 
 	data.buffer = sizeof(short)*8;
-	name.buffer = (int)"short";
+	name.buffer = (long)"short";
 	name.length = strlen((char*)name.buffer);
 	store( name );
 
 	data.buffer = sizeof(char)*8;
-	name.buffer = (int)"byte";
+	name.buffer = (long)"byte";
 	name.length = strlen((char*)name.buffer);
 	store( name );
 
 	data.buffer = 1;
-	name.buffer = (int)"bit";
+	name.buffer = (long)"bit";
 	name.length = strlen((char*)name.buffer);
 	store( name );
 
-	data.buffer = (int)&int_ind;
+	data.buffer = (long)&int_ind;
 	data.length = -1;
 	data.type = 32;
-	name.buffer = (int)"ind";
+	name.buffer = (long)"ind";
 	name.length = strlen((char*)name.buffer);
 	store( name );
 
-	data.buffer = (int)&dbl_rnd;
+	data.buffer = (long)&dbl_rnd;
 	data.length = -1;
 	data.type = 66;
-	name.buffer = (int)"rnd";
+	name.buffer = (long)"rnd";
 	name.length = strlen((char*)name.buffer);
 	store( name );
 
 	data.buffer = 0;
 	data.length = 0;
 	data.type = 0;
-	name.buffer = (int)"end";
+	name.buffer = (long)"end";
 	name.length = 3;
 	store( name );
 #ifdef GL
 	data.buffer = GL_POINTS;
 	data.type = 32;
 	data.length = 0;
-	name.buffer = (int)"GL_POINTS";
+	name.buffer = (long)"GL_POINTS";
 	name.type = 8;
 	name.length = strlen((char*)name.buffer);
 	store( name );
@@ -169,7 +169,7 @@ void init() {
 	data.buffer = GL_LINE_STRIP;
 	data.type = 32;
 	data.length = 0;
-	name.buffer = (int)"GL_LINE_STRIP";
+	name.buffer = (long)"GL_LINE_STRIP";
 	name.type = 8;
 	name.length = strlen((char*)name.buffer);
 	store( name );
@@ -177,7 +177,7 @@ void init() {
 	data.buffer = GL_LINE_LOOP;
 	data.type = 32;
 	data.length = 0;
-	name.buffer = (int)"GL_LINE_LOOP";
+	name.buffer = (long)"GL_LINE_LOOP";
 	name.type = 8;
 	name.length = strlen((char*)name.buffer);
 	store( name );
@@ -185,7 +185,7 @@ void init() {
 	data.buffer = GL_TRIANGLE_STRIP;
 	data.type = 32;
 	data.length = 0;
-	name.buffer = (int)"GL_TRIANGLE_STRIP";
+	name.buffer = (long)"GL_TRIANGLE_STRIP";
 	name.type = 8;
 	name.length = strlen((char*)name.buffer);
 	store( name );
@@ -193,48 +193,48 @@ void init() {
 	data.buffer = GL_TRIANGLE_FAN;
 	data.type = 32;
 	data.length = 0;
-	name.buffer = (int)"GL_TRIANGLE_FAN";
+	name.buffer = (long)"GL_TRIANGLE_FAN";
 	name.type = 8;
 	name.length = strlen((char*)name.buffer);
 	store( name );
 
 	data.buffer = GL_QUADS;
-	name.buffer = (int)"GL_QUADS";
+	name.buffer = (long)"GL_QUADS";
 	name.length = strlen((char*)name.buffer);
 	store( name );
 
 	data.buffer = GL_QUAD_STRIP;
-	name.buffer = (int)"GL_QUAD_STRIP";
+	name.buffer = (long)"GL_QUAD_STRIP";
 	name.length = strlen((char*)name.buffer);
 	store( name );
 
 	data.buffer = GL_V2F;
-	name.buffer = (int)"GL_V2F";
+	name.buffer = (long)"GL_V2F";
 	name.length = strlen((char*)name.buffer);
 	store( name );
 
 	data.buffer = GL_V3F;
-	name.buffer = (int)"GL_V3F";
+	name.buffer = (long)"GL_V3F";
 	name.length = strlen((char*)name.buffer);
 	store( name );
 
 	data.buffer = GL_C3F_V3F;
-	name.buffer = (int)"GL_C3F_V3F";
+	name.buffer = (long)"GL_C3F_V3F";
 	name.length = strlen((char*)name.buffer);
 	store( name );
 
 	data.buffer = GL_T2F_V3F;
-	name.buffer = (int)"GL_T2F_V3F";
+	name.buffer = (long)"GL_T2F_V3F";
 	name.length = strlen((char*)name.buffer);
 	store( name );
 
 	data.buffer = GL_PROJECTION_MATRIX;
-	name.buffer = (int)"GL_PROJECTION_MATRIX";
+	name.buffer = (long)"GL_PROJECTION_MATRIX";
 	name.length = strlen((char*)name.buffer);
 	store( name );
 
 	data.buffer = GL_MODELVIEW_MATRIX;
-	name.buffer = (int)"GL_MODELVIEW_MATRIX";
+	name.buffer = (long)"GL_MODELVIEW_MATRIX";
 	name.length = strlen((char*)name.buffer);
 	store( name );
 #endif
@@ -263,7 +263,7 @@ int main( int argc, char** argv ) {
 		simlab line;
 		line.type = 8;
 		for( int i = 1; i < argc; i++ ) {
-			line.buffer = (int)argv[i];
+			line.buffer = (long)argv[i];
 			line.length = strlen(argv[i]);
 			cmd( line );
 		}
